Changes to hydration
This is a WIP, but changes hydration so that it returns a new type,
     called a Link (could use a better name).

A link includes the rendered document fragment (called the tree) and
also contains all of the bindings, so that when an element is
disconnected its bindings can be removed.

```js
link.detach();
```

Will prevent memory links. Internally `Bram.Element` and
`Bram(SomeElement)` will do all of this for you, of course.