Change binding syntax to ${}
This does two things:

1. Changes the binding syntax from `{{foo}}` to `${foo}`. The main
reason for this is that it is familiar now that JavaScript template
literals use this syntax. And the older binding syntax has some negative
feelings among people who disliked traditional templating languages for
various reasons.

2. Removes the need for using binding syntax in `if` and `each`. So
instead of `if="{{condition}}"` you can just do `if="condition"`.

The ultimate goal is to make everything as HTML-first as possible,
    removing special syntax whenver possible. To that end we'll continue
    what the `on-` event handlers do; remove special characters and go
    with conventions.

Closes #46